[REBUILD/UPGRADE REQUIRED] - Add remote_key and remote_data to modUser
Add remote_key and remote_data fields to users table (modUser class) for maintaining remote user mappings and caching remote provider data locally. This would be for storing attributes from the remote user provider (remote_data could be accessed in PHP as an array and stored in the database as JSON to keep it compact and searchable).