Add missing modCategoryClosure to install/upgrade
Setup was not creating the table for the modCategoryClosure class during new installs of 2.0.0-pl and older. It was only added automatically in upgrades to 2.0.0-rc2 or later, so only users you upgraded to rc2 or later from releases previous to rc2 would have the table created.

This resulted in various errors during installation of certain packages.