[#2404] Allow xPDOTransport::install to be aborted when an xPDOVehicle fails to install
xPDOTransport::ABORT_INSTALL_ON_VEHICLE_FAIL attribute can now be set on a specific xPDOVehicle and if that attribute is true, the entire install process will immediately abort.

This is not the complete implementation for #2404 which more specifically asks for package-wide validators/resolvers, but it allows an initial xPDOVehicle to be used essentially as a package-wide validator.