[#2407] Add class_key criteria when requesting derived class
Automatically add a class_key criteria filter when requesting a derived class in a single-table inheritance structure, properly limiting the results to instances that are of that class explicitly.

Add xPDO::addDerivativeCriteria() method and make use of it in the three primary loader class methods, load(), loadCollection() and loadCollectionGraph(). It will be important to make sure any custom loader classes are updated with a call to addDerivativeCriteria() to get this important fix.