Fix mkdir issue in xPDOGenerator::parseSchema()
When outputting classes and maps, if the directory immediately above where the model packages starts placing files does not exist and is not writable, parseSchema will fail due to the use of mkdir() on the path. A switch to using xPDOCacheManager->writeTree() should resolve this or at least provide more options to work around file permission issues in particular environments.