[#2609] Remove filter commands and modifiers from scriptProperties
modElement and modTag instances were storing filter_commands and filter_modifiers in the properties used in the process() method. This made those available as properties in the scriptProperties array passed to an Element instance. This is not a problem unless you pass the scriptProperties to another Element. This results in recursive execution of the filters (or placeholders being set for the filter_commands/filter_modifiers properties) in any Element that was processed manually by a Snippet or Plugin that passed the scriptProperties on to that Element.

This is resolved by moving the commands and modifiers onto the modInputFilter instance and providing easier access to the filter object in modElement and modTag. See modElement/modTag::getInputFilter() and getOutputFilter(), as well as modInputFilter::hasCommands(), getCommands(), and getModifiers().