Add xPDO::getIterator() allowing more efficient traversing of result sets
You can now traverse (with foreach) an xPDOIterator instance representing a result set from an xPDOCriteria instance, and instantiate one xPDOObject instance at a time within the loop. It works essentially the same as xPDO::getCollection(), but instead of using the memory required to load the entire collection of objects at once, you only use the memory required for a single instance as you iterate over it. Memory usage is greatly reduced, and performance is slightly increased.

Just call getIterator() instead of getCollection(); the only difference is the key will be the raw sequence with getIterator (instead of the primary key as in getCollection).