[#2666] Prevent sendRedirect() from preserving request parameters unless specified
Refactored modResponse::sendRedirect() and modX::sendRedirect() method signatures to allow the second parameter to be an array of options for the method. The count_attempts, type, and responseCode parameters are now automatically turned into an options array if the second parameter is not an array to support calls with legacy parameters.

If you want sendRedirect() to preserve the referring pages REQUEST parameters, just pass a 'preserve_request' => true in your options array, e.g.

	$modx->sendRedirect('http://domain.tld', array(
		'preserve_request' => true,
		'responseCode' => 'HTTP/1.1 301 Moved Permanently'
	));

type and responseCode can be supplied in the options array OR in the legacy parameters, but if supplied in both, the values in the options array will take precedence.