[#84] Make distribution name available in manager
distro is available in version.inc.php and settings_distro was added to System Settings. This information is used to display the distribution name in the manager header, in system info, and in the full_appname data in version.inc.php trimmed of the @ characters used for filtering in the build.

Example full_appname: MODx Revolution 2.0.5-dev (git)

Also changed @svn@ to @git@ to avoid confusion since we switched to git a while back.