[#3972] fix properties not being passed to plugins
When a propertyset was tied to a Plugin's event, it's properties were
not being passed in the $scriptProperties. This was due to an incorrect
check of the $pluginPropset variable for the position of a : character.

This fix can be verified by creating a propertyset with some test properties
and attaching it to a Plugin's event, then checking the $scriptProperties
passed to the Plugin on that event.