[#4399] work around IE js error
IE8 is having troubles with event handlers staying in memory on some
elements, in particular the element that is used for the status message,
when an Ext.Fx is applied via the .pause() method.
This commit just removes the use of .pause() for IE8 and instead applies
a duration to the ghost() effect.