[#5042] Prevent additional columns defined in derivative classes being marked lazy
When using single-table inheritance via a class_key field in your object definition, if you define additional fields in a derivative object element, when the object is retrieved via the base class rather than the derivative, those additional fields are not in the result set and thus were being marked as lazy which prevented the derivative instances from being save()'d. This change only marks fields lazy when loaded from a result set based on the fields defined for the requested class name, not the actual derivative class name defined in a class_key column.