[#5556] Enable connection pooling with master/slave support
Made possible via xPDO implementation of feature #4590.

This currently requires manual configuration via $config_options in the primary MODX config file.

Example:

    $config_options = array (
        xPDO::OPT_CONNECTIONS => array(
            array(
                'dsn' => 'mysql:host=127.0.0.1:19571;dbname=revo_replication;charset=utf8',
                'username' => 'user',
                'password' => 'pwd',
                'options' => array(xPDO::OPT_CONN_MUTABLE => false),
                'driverOptions' => array(),
            ),
            array(
                'dsn' => 'mysql:host=127.0.0.1:19572;dbname=revo_replication;charset=utf8',
                'username' => 'user',
                'password' => 'password',
                'options' => array(xPDO::OPT_CONN_MUTABLE => false),
                'driverOptions' => array(),
            ),
        ),
        xPDO::OPT_CONN_MUTABLE => true,
        xPDO::OPT_CONN_INIT => array(xPDO::OPT_CONN_MUTABLE => false)
    );

xPDO::OPT_CONN_INIT defines what kind of connection should be established when modX is instantiated; in the example, xPDO::OPT_CONN_MUTABLE is false which means a read-only connection is established.

Overrides passed to the modX constructor are used to force mutable connections in the manager and connector index.php files (as well as manager/min/index.php); this is to avoid manager integrity issues when getting random connections to various slaves that may not be immediately up-to-date.