Minor update to description of TV Value Description
As a MODX n00b I was unsure what to use as a seperator between values for TVs that can have multuple option values such as a Listbox or Radio Options.  I was surprised that the seperator was not shown in the description below, this fixes that and removes the ambiguity when adding option values.