Avoid install errors from modContextSetting vehicles
When UPDATE_OBJECT is true and an xPDOObjectVehicle payload with a compound primary key is installed without UNIQUE_KEY specified, an attempt is made to update the values from the selection criteria, which is invalid if the NATIVE_KEY is used, e.g. array(0 => 'field1', 1 => 'field2'). This registers errors like the following in the install error logs that are generated during upgrades:

(ERROR @ /revolution/setup/index.php) xPDOObject - Called set() with an invalid field name: 0
(ERROR @ /revolution/setup/index.php) xPDOObject - Called set() with an invalid field name: 1