Set allow_tags_in_post false in default System Settings
This should ONLY be true in Context Settings for mgr or any Contexts that need to provide the ability for users to POST MODX tags, numeric entities or script tags in requests to those Contexts.

Also fixes the description which erroneously indicated that when false, HTML tags were stripped from the POST. Only HTML script tags, numeric entities, and MODX tags are stripped from the POST when this is false. NOTE that this is ALWAYS done to $_REQUEST, $_GET, and $_COOKIE vars.