Minor Header CSS & A crappy way to position footer
I'm flabbergasted that both #modx-content and #modx-leftbar are
position:absolute. It completely does away with making any use of the
box model. I almost had everything laid out without either being
absolute, and using floats, but them needing to both be 100% doesn't
allow for anything to go under (especially with body overflow:hidden)
so I came up with this for the footer for now which I don't love