Additional method that allows easy saving of array elements as .csv file (fputcsv())
I don't know if you find this a helpful addition or not, just thought I share if so. I really like the modFileHandler and use it a lot and this method is something I find quite convenient when generating an export of any data that can be stored in an array (ie . Collections of MODx objects or whatever objects etc.). Actually it involves a bit more than just this method (need to generate headers, catch encoding issues etc.), but when the data is prepared nicely in an array, this just needs to be passed to this method and voila...