Sass Switch: READ ME
If you are going to be contributing to themes you must read this.

templates/default/sass/index.scss is now used to build the index.css
file. It prepends includes xtheme-modx.scss and will be minified as
well.

DO NOT COMMIT OR MAKE CHANGES to index.css or login.css. They will be
overwritten. Use the respective Sass files.

Workflow for this is still being ironed out, let us know if you need
help getting up and running with Sass.