Disabled drag and drop on the resource tree if the sort order not is menu index.
This will prevent MODX to save the entiry tree in a different order.

Example:
I have hundreds of resources and i sort the tree on published on. The
tree isn’t sorted on menu index anymore and i accidently dd a resource
the menu index of the entire site is changed.