Bring back backwards compatibility for the help window.
In 2.2, overwriting MODx.config.help_url worked for loading a custom help url on CMPs. With this patch, that is once again possible. It prefers the new method (MODx.helpUrl, coming from the modManagerController.getHelpUrl() method) but checks the MODx.config.help_url for backwards compatibility.

It also prepends the base_help_url if the help url does not start with a protocol. Before this patch that was done on the server side (and stored in MODx.helpUrl), however that meant losing the ability to dynamically changing the help url in the manager. Putting that logic in the loadHelpPane method instead has the same result, but is more flexible.