Update modx.grid.js
When you enable grouping on a MODx.grid.Grid, it will apply a Ext.grid.GroupingView to the grid, which is great feature, but as I want to hide/disable the "group by this column" feature when you click on the little arrow near each column header, I need to pass "enableGroupingMenu: false" to the GroupingView. But I cannot do that (whithout touching the core files).

Maybe good to have a "groupingConfig: { }" which will be passed when you set this up in your grid. This is how it works for me.