Set default background color for thumbs to WHITE instead of BLACK
The default background color for thumbs of "PNGs with transparency" used to be black, which does not play well together with the white background of the resource editing page. BLACK is never the right choice.
Another solution would have been to change the default filetype $f="png", but this would lead to bigger thumbnail files.