Fiting portrait and landscape imageTV thumbs into a maximum area of 400x400px
### Summary
This is an improvement for #12491 which only limited the the width to 400px, the height remained unlimited.
Now the script also handles very tall images like skyscraper banners (120x600px) and fits them into a 400x400px area. This saves a lot of vertical space for portrait format images.

### Step to reproduce
Just select an image for a TV.

### Observed behavior
The maximum thumbnail size is now fitted into 400x400px.

### Expected behavior
Just as observed. ;-)