Add anonymous_sessions to allow session-less access for anonymous users
Default behavior is unchanged and set to true. When disabled however, anonymous users will not start a session when accessing a page on the site; only logged in users will have access to a session.

Also adds modX->startSession() which starts a session if not already and simply returns true if a session is available, regardless how it was started