Implement X-Powered-By header to send "MODX Revolution 2.x" on all requests
This header will allow MODX usage to be more easily tracked by third party tools. As there is a slight enhanced security risk to sites announcing what it is built with, and this hasn't been done historically, this setting defaults to being off and requires to be opted in to.

The opt-in is only presented on new installs, so on upgrades the setting is not changed. The default value is off.

 Based on the discussion in #12882