Prevent reflected XSS in connector's JSONP support
This is a low risk security issue. It allows a client (browser) to potentially execute javascript when a specifically forged request was present. If an attacker is able of forging such a request, they already have access to the client in a way that renders XSS attempts like these fairly.. well.. pointless. Nevertheless, can't hurt to make sure it can't be abused anyway.

Reported by Kacper Szurek in security ticket 28 on May 29th