Duplicate array index, and description is missing
The second $_lang['setting_anonymous_sessions'] should be $_lang['setting_anonymous_sessions_desc'] with a different text. While this is correct in the 2.x branch, it isn't in the 2.5.x branch for some reason, and the language file delivered with MODX 2.5.1-pl contains the wrong version, too. This is why I corrected it this way, not via crowdin (which is impossible).