package.grid.js // default_per_page items in Package-Overview
### What does it do?
Updated package.grid.js, so that it not statically always show 10 entries, 
but acknowledges the _default_per_page_ system-setting the user provided, 
if it is larger than 10.

### Why is it needed?
The Package-Overview statically shows 10 entries. 
I was expecting it to behave as all the other grids in the manager.

### Related issue(s)/PR(s)
None that I know of.