Possible fix for blind SQL injection
For now, if there is a string when you get an object, this string will be got as an criteria for SQL query. So, if you want just ti get and object with primary key, but not filter a user request enough, there could be a blind SQL injection.

This fix will force all conditions to be an:
1. primary key
2. array with keys and values for query
3. or xPDOQuery

So no RAW SQL queries could be made with `getObject` call.