SQL injections in ORDER BY and LIMIT clauses
Following a report from Nikolay Lanets on November 4th, several SQL injections in xPDO have been found and fixed.

When creating a SQL query through an xPDOQuery object (as virtually any core and third party code interacting with the database does), the `sortby` and `limit` methods accepted arbitrary SQL which was not properly sanitised.

While this vulnerability existed within the xPDOQuery object, the original report from Nikolay and further testing shows that it could be exploited without writing and executing arbitrary code. This includes by providing specially crafted arguements to connectors/processors (requiring a valid manager session), or by a snippet call that passes a sort or limit to xPDOQuery directly.

As these are common scenarios that can't reasonably be resolved in each specific calling function, this has been patched at the xPDOQuery level at the cost of a breaking change to extras (and possibly MODX core code) that do rely on the sort accepting abitrary SQL, for example the common approach for sorting resources in a specific order with sort clauses like `FIELD(modResource.id, 1, 2, 3)`. These order clauses will no longer work following this patch, unless the snippet indicates the input should be trusted by calling the `sortbyRaw` method with a value of `true`.