Issue 13323, #modxbughunt
If you are moving textfields and textareas to

modx-resource-main-left
or
modx-resource-main-right

they are slightly too wide on the right side, breaking the boundaries of their surrounding columns.

Width:99% combined with padding: 5px may result in containers bigger than 100%. We can use box-sizing: border-box to prevent this behaviour. 

Box-sizonh is supported by 98% of all browsers: http://caniuse.com/#search=box-sizing Even if it is not supported, the form fields should still be usable.