Change getResourceURI logic
I have an issue near #13278. I prefer to get copy of document in container, beacause it is shorter way to publish new document (you dont need to choose template, what field is used etc). With 2.5.5 when i make a copy of a document i get error message like this:

(WARN @ ...\core\model\modx\modcontext.class.php : 244) `126` was requested but no alias was located.
(ERROR in modContext::makeUrl @ ..\core\model\modx\modcontext.class.php : 321) Resource with id 126 was not found in context web

I make some investigation and detect, what on line 241 we check if document has alias by executing getResourceURI($id). If this method return null, we think that document not exist and dont change $found variable to true. When we did not do this, we have this error message in logs. But this is not true, document existing and have alias.

Continuing investigation i found, what the logic of getResourceURI method is check if cache_alias_map is set to true and if aliasMap is generated otherwise method request document from database. But wit my case (when i copy document) aliasMap is formed, but document not add in it. I think where is two solution of it: 1. Theoretically use plugin what will renewing aliasMap after coping document. 2. Change the logic of getResourceURI to be sure, what document really don't exist or dont have uri (this situation is near error for me). 

This solution will not give alot of additional load, because if aliasMap is formed and contains document inside the database execution will be skipped. This solution tested by me and work fine.