Add OPTIONS request method Used in cross-origin HTTP request - https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS - http://zacstewart.com/2012/04/14/http-options-method.html
An additional class property $allowedMethod is introduced to be filled with an array of allowed request methods.
Another additional class properties $allowedMethod could be filled with an array of allowed request headers.

I am not sure about allowing the OPTIONS request method in authentificated controllers, so this part in the modRestService class could be removed.