Prevent reload of login page
If you can't login to manager you will see an error, but the next attempt will just reload window. It is because user is actually logged it but can't see backend, so manager controller will check it and logout him on the next request. That is why we have a page reload instead of new attempt to login.

This change will fix it.