Added a new property fallbackResponse to Ext.dataConnection: This object emulates the structure of the json object retuend by the server.
Added a new function isValidResponse() tp Ext.data.Connection:
This function tests whether the passed in response object's responseText is valid JSON.
It returns the original stringified object if  valid.
It returns the stringified version fallbackResponse object if the result is false.

The message property of fallbackResponse is populated with an error message that should be replaced with a new entry in the lexicon