Remove cache_system_settings from the installation
This setting cannot logically work. When MODX is initialized, it has to first load the settings (which will use the cache) before it can determine if the cache should be used. Does not compute. If people want to disable the system settings cache, which I strongly recommend against, it has to be added to the $config_options array in the core/config/config.inc.php file.