- Ability to customize dashboard: move, resize, change order and size of widgets. - Improved look and feel - 2 new widgets: Buttons and Updates - Dashboard and widgets use so minimum of ExtJS as possible - Added new column "customizable" to modDashboard - Added 2 new columns to modDashboardWidget: "properties" and "permission" - Added 2 new columns to modDashboardWidgetPlacement: "user" and "size" - Added new aggregate to modUser: "DashboardWidgets" - Default dashboard and widgets will not be overwritten on upgrade anymore - Recently edited resources widget now shows all users and requires "view_document" permission
Dashboard of a user can be custimizable, or not. If it is, user can add new widgets, remove old, resize, change size and order by drag-n-drop.
If Dashboard is not customizable, it will be static as admin specified it to be. User will not be able to do anything, as it was before.

Widgets now have JSON properties similar to modUserProfile extended field, that could be used by developers for any puprose. For example, you can change all links in texts in default Buttons widget.
Also widgets can now require any permission to install them on users Dashboard.