Restore html in resource tree (#14358) while preserving XSS protections in trees by default
This introduces a new renderItemText function that trees can define to
determine how their text field should be rendered. When not set, the
htmlEncode formatter is used.

This new function is added to the resource and menu trees. Those elements
already have XSS protection server-side, where the user input is escaped
and the resulting text contains HTML that the system expects to be
rendered.