Fix xPDOFileCache->delete() when a directory with same cache key exists
Other xPDOCache implementations check for multiple_object_delete option before attempting to recursively remove cache entries that match the cache key provided to the delete() method. xPDOFileCache was not doing this, and further, if a directory existed with the cache key to be deleted, the directory was emptied of cache entries but the actual cache entry that was meant to be deleted was left in place.

For 2.x, cache entries in a folder matching the cache key will be removed recursively unless multiple_object_delete is explicitly set to false. This is to prevent a major change in behavior. For 3.x, it will be made consistent with other xPDOCache implementations and require multiple_object_delete to be explicitly set to true before entries in a folder matching the cache key will be removed.

Resolves #170