Proposed Change to TV Creation in Manager (Initial)
Introduces a process that incorporates TV input and output options fields into the form as individual Ext components instead of as a rendered DOM element. This allows for true, full interaction with and validation of each field. Note, changes to the output options fields to come in next commit.

Other features, such as tv type-specific default value fields and labels are included.

These changes also explore a different, hopefully more reliable way of assessing and reacting to a form's dirty state.