Proposal to fix security issue reported Sept 2nd
This incorporates the `view` permission as required for interaction with a media source. The `load` permission is hardcoded as to always be available, as a media source must be available for front-end usage. The primary concern is that neither a user in the manager nor a programmatic interaction with the media source should inadvertantly gain access to a source it is not intended to access.

Currently, there are 2 checks on the modMediaSource object that are routinely checked for their return in programmatic interaction. Either getWorkingContext(), or initialize(). By implementing a checkPolicy in both of those method, and adjusting derivative classes accordingly, it is expected that all programmatic access to media sources (including processors and third party extras) are sufficiently protected with this check.

The Create and Edit file controllers are also adjusted to specifically check $source->initialize() and to return the appropriate failure when not present.