Fixed display of the pagination toolbar
Fixed bug with display of the pagination toolbar. If the request to the processor was delayed, the `expand` function had time to work because of this piece of code:
```
if (this.mode == 'remote' && !this.loaded && this.tries < 4) {
    // Store not yet loaded, let's wait a little bit
    this.tries += 1;
    Ext.defer(this.expand, 250, this);
    return false;
}
```
And when the answer came from the processor, the `expand' function was not started anymore. We had to make a small crutch that instantly collapse the list and expanded it again.