Add `upload_check_exists` system setting to allow overwriting files with same name on upload + fix self-xss with payload in the file name
As discussed in https://community.modx.com/t/version-2-8-is-out-easy-replacement-of-files-pdf-images-is-axed/3166 and the discussion in #15232, some people are used to being able of replacing files simply by uploading them with the same name. That behavior was removed in #15232 (v2.8.0-pl) to prefer a stricter check where users are presented with an error when a file already exists with the same name.

The new setting, enabled by default, allows users to return to that old behavior and bypassing the check on upload.

(The check is still in use, and valid, when renaming files. This is because the renameObject method on the media source will also fail if the target file already exists, so it makes sense to keep that there with the nice error.)

Also noticed the `file_err_ae` error message used the unescaped objectname for rendering the error, which could lead to a self-xss. Tidied that up.