Restrict file management to allowed file types [SEC-3837]
While creating/uploading files to a media source was properly limited to the allowed file types (either media source allowedFileTypes, or the upload_* settings), this was not the case for various other actions and most notably updating files. In security report 3837 this was reported as a potential RCE by editing PHP files through manager XSS attacks. Personally I find labelling that a RCE debatable considering editing files is a core feature, not a bug, but nevertheless it's definitely a bug that it allowed editing non-allowed file types.